Refactor prototyp_linnea, modify .gitignore and remove excess files
