+++ .gitignore
@@ -1,3 +1,4 @@
 venv/
 *.pyc
 .idea/
+__pycache__/
