+++ components/reset_and_search/Reset_and_search.py
